+++ src/models.rs
@@ -2,7 +2,7 @@
 pub struct Expense {
   pub user_id: i64,
   pub username: String,
-  pub price: f64,
+  pub price: f32,
   pub comments: String,
   pub date: String,
 }
